Extract Esc to dimiss all toasts into custom hook
